+++ src/components/layout/Main/styles.ts
